refactor(sidebar): extract previews to keep SidebarView under file_length
Adding a second SidebarView #Preview in the previous commit pushed
SidebarView.swift past SwiftLint's 400-line file_length budget. Move
both previews and the seedSidebarPreview(backend:) helper into a new
sibling file SidebarView+Previews.swift; xcodegen auto-discovers the
new file, so no project.yml change is needed. SidebarView.swift drops
back under the budget without any compensating shrinks to the
production code.
